services/profile_service/requirements.txt
# FastAPI и веб-сервер
fastapi==0.115.9
uvicorn==0.34.0
starlette==0.45.3

# База данных и ORM
SQLAlchemy==2.0.38
alembic==1.14.1
psycopg2-binary==2.9.10
psycopg
psycopg-binary

# Redis для кэширования
redis==5.0.1

# Pydantic для схем и валидации
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.8.1

# Конфигурация
python-dotenv==1.0.1

# Для обработки форм и файлов
python-multipart==0.0.6

# Валидация email
email_validator==2.2.0

# HTTP клиент (для межсервисного взаимодействия)
httpx==0.25.2
requests==2.32.3

# JWT для локальной валидации токенов (опционально)
python-jose[cryptography]==3.4.0

# Для работы с датами и временем
python-dateutil==2.9.0

# Для работы с изображениями и аватарами
Pillow==10.4.0

# Для асинхронной работы с файлами
aiofiles==24.1.0

# Зависимости системы
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
jose==1.0.0
Mako==1.3.9
MarkupSafe==3.0.2
pyasn1==0.4.8
pycparser==2.22
rsa==4.9
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0

# Тестирование (для разработки)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0